test(spp): split circuit_test.go monolith by theme
Splits the ~900-line circuit_test.go into focused same-package files: shared
fixtures/helpers in helpers_test.go and tests grouped by concern (merkle path
inclusion/non-inclusion, nullifier/double-spend, owner+P256+signature, balance
+ public asset movement). No test bodies changed — the full set of functions is
identical, only relocated.
